Refactor AnalysisPrinter part1
